[cpp] move each entity to separate file
