+++ cpp/microbench/workloads/args_generators/args_generator_json_convector.h
@@ -5,15 +5,15 @@
 
 #include "nlohmann/json.hpp"
 #include "args_generator_builder.h"
-#include "workloads/args_generators/impls/default_args_generator.h"
-#include "workloads/args_generators/impls/skewed_sets_args_generator.h"
-#include "workloads/args_generators/impls/skewed_insert_args_generator.h"
-#include "workloads/args_generators/impls/creakers_and_wave_args_generator.h"
-#include "workloads/args_generators/impls/temporary_skewed_args_generator.h"
-#include "workloads/args_generators/impls/leafs_handshake_args_generator.h"
-#include "workloads/args_generators/impls/generalized_args_generator.h"
-#include "workloads/args_generators/impls/null_args_generator.h"
-#include "workloads/args_generators/impls/range_query_args_generator.h"
+#include "workloads/args_generators/impls/default_args_generator_builder.h"
+#include "workloads/args_generators/impls/skewed_sets_args_generator_builder.h"
+#include "workloads/args_generators/impls/skewed_insert_args_generator_builder.h"
+#include "workloads/args_generators/impls/creakers_and_wave_args_generator_builder.h"
+#include "workloads/args_generators/impls/temporary_skewed_args_generator_builder.h"
+#include "workloads/args_generators/impls/leafs_handshake_args_generator_builder.h"
+#include "workloads/args_generators/impls/generalized_args_generator_builder.h"
+#include "workloads/args_generators/impls/null_args_generator_builder.h"
+#include "workloads/args_generators/impls/range_query_args_generator_builder.h"
 #include "errors.h"
 
 namespace microbench::workload {
